Support for Duplicate Projects!
Using the _id field as the sole indexer for each Project. and using the generated ObjectID stored as a string as the ProjectID to be query'd for throughout the rest of the Project's existence and saving this id string in the projects pertaining to each user in the User Collection and projects pertaining to each Tag which corresponds to the project in the Project Collection.